requirements.txt
google-generativeai>=0.8.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
Flask>=3.1.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.0
Flask-Mail>=0.9.1
Werkzeug>=3.0.0
email-validator>=2.1.0
bcrypt>=4.1.0
psycopg2-binary>=2.9.9
reportlab>=4.0.0
matplotlib>=3.8.0
wikipedia>=1.4.0